\name{bintanja05}
\alias{bintanja05}
\docType{data}
\title{bintanja05 palaeoclimate recontsruction}
\description{
Inverse modelling recontruction of ice-sheet contribution
to the varabilitiy in oxygen isotope composition, eastatic
sea-level and deep-ocean temperatue, based on the 
Lisiecki and Raymo benthic stack

Citation: Bintanja et al. (2005)

}
\usage{data(bintanja05)}
\format{
  A data frame with the following  variables.
  \describe{
\item{\code{1 }}{Time (kyr BP)}
\item{\code{1 }}{Atmospheric surface air temperature relative to present (degC)}
\item{\code{1 }}{Global sea level relative to present (m)}
\item{\code{1 }}{Eurasian ice volume relative to present (m sea level equivalent)}
\item{\code{1 }}{North American ice volume relative to present (m sea level equivalent)}
\item{\code{1 }}{Marine oxygen isotope value relative to present (o/oo)}
\item{\code{1 }}{Ice sheet contribution to the marine isotope signal, relative to present (o/oo)}
\item{\code{1 }}{Deep-sea temperature contribution to the marine isotope signal, relative to present (o/oo)}
\item{\code{1 }}{Mean isotope value of the Eurasian ice sheets (o/oo)}
\item{\code{1 }}{Mean isotope value of the North American ice sheets (o/oo)}
}
}
\source{
\url{ftp://ftp.ncdc.noaa.gov/pub/data/paleo/contributions_by_author/bintanja2005/bintanja2005.txt}
}

\references{
Bintanja, R., R.S.W. van de Wal, and J. Oerlemans. 2005. 
Modelled atmospheric temperatures and global sea levels 
over the past million years. 
Nature, Vol. 437, pp. 125-128, 1 September 2005. 
doi:10.1038/nature03975 
}

\examples{
data(bintanja05)
}
\keyword{datasets}
